Add a wireguard network with hades as a router
Hades is the only host in the hetzner private network that needs to have
wireguard installed. The others route packets to hades due to a routing
rule set in hetzner. Hades forwards packets between the wireguard subnet
(10.83.1.0/24) and the hetzner subnet (10.83.0.0/24). Currently one peer
has been set up, mjukglass, but _Arkaden_ in META is apparently behind a
NAT so we could add that as well so people can SSH to it (probably
through a proxy (probably traefik) so we don't need to add people to the
wireguard network).