remove tofu apply GitHub workflow
Although interesting in theory, this was just a hindrance in practice,
since it locks the state for every push and we very seldom use pull
requests anyway (so there's no pre-prepared plan to apply).

It might be worth considering adding it back in the future if the
conditions so facilitate it.